API

API

DaskExecutor

DaskDistributedDispatcher.DaskExecutor — Type.

DaskExecutor is an Executor which executes julia computations as scheduled by the python dask-scheduler. It can run computations both asynchronously or in parallel (if Workers are started on a julia cluster instead).

DaskExecutor's dispatch!(::DaskExecutor, ::Dispatcher.DispatchNode) method will complete as long as there are no cycles in the computation graph, the dask-scheduler remains online, and there is at least one Worker that is listening to the dask-scheduler.

source

DaskDistributedDispatcher.DaskExecutor — Method.

DaskExecutor(scheduler_address::String="$(getipaddr()):8786")

Return a new DaskExecutor. The scheduler_address only needs to be included if the dask-scheduler is running on a different machine or not on it's default port (8786).

NOTE: A dask-scheduler must be running at all times or the DaskExecutor execution will fail. If the scheduler is taken offline during execution for some reason, any remaining operations will fail to complete. Start a dask-scheduler from a terminal by typing dask-scheduler:

$ dask-scheduler
Start scheduler at 192.168.0.1:8786

Prerequisites

python 2.7 or 3.5

the python dask.distributed package (instructions for install here)

Note that using the dask-scheduler and executing compuations in a distributed manner can add overhead for simple tasks. Consider using an AsyncExecuter or ParallelExecuter if possible. The advantage that using the dask-scheduler has is that it schedules computations in a manner that is short-term-efficient and long-term-fair.

Usage

The DaskExecutor can run both asynchronously with the Workers, or in parallel if Workers are spawned on separate julia processes in a cluster.

NOTE: Users must startup at least one Worker by pointing it to the dask-scheduler's address or else run! will hang indefinetely.

Examples

Running asynchronously:

# Reminder: make sure the dask-scheduler is running
using DaskDistributedDispatcher
using Dispatcher
using ResultTypes

Worker()

exec = DaskExecutor()
ctx = DispatchContext()
n1 = add!(ctx, Op(()->3))
n2 = add!(ctx, Op(()->4))

results = run!(exec, ctx)

fetch(unwrap(results[1]))  # 3
fetch(unwrap(results[2]))  # 4

Running in parallel:

# Reminder: make sure the dask-scheduler is running
using DaskDistributedDispatcher
using Dispatcher
using ResultTypes

addprocs(3)
@everywhere using DaskDistributedDispatcher

for i in 1:3
    cond = @spawn Worker()
    wait(cond)
end

exec = DaskExecutor()
ctx = DispatchContext()
n1 = add!(ctx, Op(()->3))
n2 = add!(ctx, Op(()->4))

results = run!(exec, ctx)

fetch(unwrap(results[1]))  # 3
fetch(unwrap(results[2]))  # 4

To delete all previously computed information from the workers:

reset(exec)

Advanced Workflows

It is possible to bypass the DaskExecutor and use the Client directly to submit compuations, cancel previously scheduled DispatchNodes, gather results, or replicate data across all workers. See Client for more details. It is recommened to start with a DaskExecutor and access its client field if needed later on.

source

Dispatcher.dispatch! — Method.

dispatch!(exec::DaskExecutor, ctx::DispatchContext; throw_error=true) -> Vector

The default dispatch! method uses asyncmap over all nodes in the context to call dispatch!(exec, node). These dispatch! calls for each node are wrapped in various retry and error handling methods.

Wrapping Details

All nodes are wrapped in a try catch which waits on the value returned from the dispatch!(exec, node) call. Any errors are caught and used to create DependencyErrors which are thrown. If no errors are produced then the node is returned.

NOTE: All errors thrown by trying to run dispatch!(exec, node) are wrapped in a DependencyError.

By default the DaskExecutor has no retry_on functions since retrying failed ops is not explicitly supported by the dask-scheduler.

A node may enter a failed state if it exits the retry wrapper with an exception. In the situation where a node has entered a failed state and the node is an Op then the op.result is set to the DependencyError, signifying the node's failure to any dependent nodes. Finally, if throw_error is true then the DependencyError will be immediately thrown in the current process without allowing other nodes to finish. If throw_error is false then the DependencyError is not thrown and it will be returned in the array of passing and failing nodes.

Arguments

exec::DaskExecutor: the executor we're running

ctx::DispatchContext: the context of nodes to run

Keyword Arguments

throw_error::Bool=true: whether or not to throw the DependencyError for failed nodes

Returns

Vector{Union{DispatchNode, DependencyError}}: a list of DispatchNodes or DependencyErrors for failed nodes

Throws

dispatch! has the same behaviour on exceptions as asyncmap and pmap. In 0.5 this will throw a CompositeException containing DependencyErrors, while in 0.6 this will simply throw the first DependencyError.

source

DaskDistributedDispatcher.reset! — Method.

reset!(exec::DaskExecutor)

Restarts the executor's Client, which tells the scheduler to delete previously computed data since it is not needed anymore. The scheduler, in turn, signals this to the workers.

source

DaskDistributedDispatcher.retries — Method.

retries(exec::DaskExecutor) -> Int

Return the number of retries per node.

source

DaskDistributedDispatcher.retry_on — Method.

retry_on(exec::DaskExecutor) -> Vector{Function}

Return the array of retry conditions.

source

DaskDistributedDispatcher.dispatch! — Method.

dispatch!(exec::DaskExecutor, node::Dispatcher.DispatchNode) -> Future

dispatch! takes the DaskExecutor and a DispatchNode to run and submits the DispatchNode to the Client for scheduling.

This is the defining method of DaskExecutor.

source

Client

DaskDistributedDispatcher.Client — Type.

Client

Client that can be interacted with to submit computations to the scheduler and gather results. Should only be used directly for advanced workflows. See DaskExecutor instead for normal usage.

Fields

nodes::Dict{String, DispatchNode}: maps keys to their dispatcher DispatchNode

id::String: this client's identifier

status::String: status of this client

scheduler_address::Address: the dask-distributed scheduler ip address and port info

scheduler::Rpc: manager for discrete send/receive open connections to the scheduler

connecting_to_scheduler::Bool: if client is currently trying to connect to the scheduler

scheduler_comm::Nullable{BatchedSend}: batched stream for communication with scheduler

pending_msg_buffer::Array: pending msgs to send on the batched stream

source

DaskDistributedDispatcher.Client — Method.

Client(scheduler_address::String) -> Client

Construct a Client which can then be used to submit computations or gather results from the dask-scheduler process.

Usage

using DaskDistributedDispatcher
using Dispatcher

addprocs(3)
@everywhere using DaskDistributedDispatcher

for i in 1:3
    @spawn Worker("127.0.0.1:8786")
end

client = Client("127.0.0.1:8786")

op = Op(Int, 2.0)
submit(client, op)
result = fetch(op)

Previously submitted Ops can be cancelled by calling:

cancel(client, [op])

# Or if using the `DaskExecutor`
cancel(executor.client, [op])

If needed, which worker(s) to run the computations on can be explicitly specified by returning the worker's address when starting a new worker:

using DaskDistributedDispatcher
client = Client("127.0.0.1:8786")

pnums = addprocs(1)
@everywhere using DaskDistributedDispatcher

worker_address = @fetchfrom pnums[1] begin
    worker = Worker("127.0.0.1:8786")
    return worker.address
end

op = Op(Int, 1.0)
submit(client, op, workers=[worker_address])
result = result(client, op)

source

DaskDistributedDispatcher.submit — Method.

submit(client::Client, node::DispatchNode; workers::Array{Address,1}=Array{Address,1}())

Submit the node computation unit to the dask-scheduler for computation. Also submits all node's dependencies to the scheduler if they have not previously been submitted.

source

DaskDistributedDispatcher.cancel — Method.

cancel{T<:DispatchNode}(client::Client, nodes::Array{T, 1})

Cancel all DispatchNodes in nodes. This stops future tasks from being scheduled if they have not yet run and deletes them if they have already run. After calling, this result and all dependent results will no longer be accessible.

source

DaskDistributedDispatcher.gather — Method.

gather{T<:DispatchNode}(client::Client, nodes::Array{T, 1})

Gather the results of all nodes. Requires there to be at least one worker available to the scheduler or hangs indefinetely waiting for the results.

source

DaskDistributedDispatcher.replicate — Method.

replicate{T<:DispatchNode}(client::Client; nodes::Array{T, 1}=DispatchNode[])

Copy data onto many workers. Helps to broadcast frequently accessed data and improve resilience.

source

DaskDistributedDispatcher.shutdown — Method.

shutdown(client::Client)

Tell the dask-scheduler that this client is shutting down. Does NOT terminate the scheduler itself nor the workers. This does not have to be called after a session but is useful to delete all the information submitted by the client from the scheduler and workers (such as between test runs). To reconnect to the scheduler after calling this function set up a new client.

source

DaskDistributedDispatcher.get_key — Method.

get_key{T<:DispatchNode}(node::T)

Calculate an identifying key for node. Keys are re-used for identical nodes to avoid unnecessary computations.

source

DaskDistributedDispatcher.ensure_connected — Method.

ensure_connected(client::Client)

Ensure the client is connected to the dask-scheduler. For internal use.

source

DaskDistributedDispatcher.send_to_scheduler — Method.

send_to_scheduler(client::Client, msg::Dict{String, Message})

Send msg to the dask-scheduler that the client is connected to. For internal use.

source

DaskDistributedDispatcher.serialize_deps — Method.

serialize_deps(client::Client, deps::Array, tkeys::Array, tasks::Dict, tasks_deps::Dict)

Serialize all dependencies in deps to send to the scheduler. For internal use.

Returns

tkeys::Array: the keys that will be sent to the scheduler in byte form

tasks::Dict: the serialized tasks that will be sent to the scheduler

tasks_deps::Dict: the keys of the dependencies for each task

source

DaskDistributedDispatcher.serialize_node — Method.

serialize_node(client::Client, node::DispatchNode)

Serialize all dependencies in deps to send to the scheduler. For internal use.

Returns a tuple of:

task::Dict: serialized task that will be sent to the scheduler

unprocessed_deps::Array: list of dependencies that haven't been serialized yet

task_dependencies::Array: the keys of the dependencies for task

source

DaskDistributedDispatcher.serialize_task — Method.

serialize_task{T<:DispatchNode}(client::Client, node::T, deps::Array) -> Dict

Serialize node into its components. For internal use.

source

Address

DaskDistributedDispatcher.Address — Type.

Address

A representation of an endpoint that can be connected to. It is categorized by its scheme (tcp is currently the only protocol supported), host, and port.

source

DaskDistributedDispatcher.Address — Method.

Address(address::String) -> Address

Parse address and returns the corresponding Address object.

source

DaskDistributedDispatcher.Address — Method.

Address(host::IPAddr, port::Integer)) -> Address

Return the corresponding Address object to the components host and port. By default the tcp protocol is assumed.

source

Base.show — Method.

show(io::IO, address::Address)

Print a representation of the address to io. The format used to represent addresses is "tcp://127.0.0.1:port".

source

Base.connect — Method.

Base.connect(address::Address)

Open a tcp connection to address.

source

MsgPack.pack — Method.

MsgPack.pack(io::Base.AbstractIOBuffer{Array{UInt8,1}}, address::Address)

Pack address as its string representation.

source

DaskDistributedDispatcher.parse_address — Method.

parse_address(address::String) -> (String, IpAddr, Integer)

Parse an address into its scheme, host, and port components.

source