API

API

Client

Client
Client(::String)
submit(::Client, ::Dispatcher.Op)
result(::Client, ::Dispatcher.Op)
cancel(::Client, ::Array)
gather(::Client, ::Array)
shutdown(::Client)
default_client()

Worker

DaskDistributedDispatcher.Worker — Type.

Worker

A Worker represents a worker endpoint in the distributed cluster that accepts instructions from the scheduler, fetches dependencies, executes compuations, stores data, and communicates state to the scheduler.

source

DaskDistributedDispatcher.Worker — Method.

Worker(scheduler_address::String)

Creates a Worker that listens on a random port for incoming messages.

source

DaskDistributedDispatcher.address — Method.

address(worker::Worker)

Returns this Workers's address formatted as an URI.

source

Base.show — Method.

show(io::IO, worker::Worker)

Prints a representation of the worker and it's state.

source