API

API

Client

DaskDistributedDispatcher.Client — Type.

Client

A Client represents a client that the user can interact with to submit computations to the scheduler and gather results.

Fields

ops::Dict{String, Dispatcher.Op}: maps keys to their dispatcher ops

id::String: this client's identifier

status::String: status of this client

scheduler_address::Address: the dask-distributed scheduler ip address and port information

scheduler::Rpc: manager for discrete send/receive open connections to the scheduler

connecting_to_scheduler::Bool: if client is currently trying to connect to the scheduler

scheduler_comm::Nullable{BatchedSend}: batched stream for communication with scheduler

pending_msg_buffer::Array: pending msgs to send on the batched stream

source

DaskDistributedDispatcher.Client — Method.

Client(scheduler_address::String; throw_errors::Bool=true) -> Client

Construct a Client which can then be used to submit computations or gather results from the dask-scheduler process. Set throw_errors to false if you want to instead receive string representations of any errors thrown while running a user submitted task.

source

DaskDistributedDispatcher.submit — Method.

submit(client::Client, op::Dispatcher.Op; workers::Array=[])

Submit the Op computation unit to the dask-scheduler for computation.

source

DaskDistributedDispatcher.result — Method.

result(client::Client, op::Dispatcher.Op) -> Any

Gather the result of the Op computation unit. Requires there to be at least one worker available to the scheduler or hangs indefinetely.

source

DaskDistributedDispatcher.cancel — Method.

cancel(client::Client, ops::Array{Dispatcher.Op})

Cancel all Ops in ops. This stops future tasks from being scheduled if they have not yet run and deletes them if they have already run. After calling, this result and all dependent results will no longer be accessible.

source

DaskDistributedDispatcher.gather — Method.

gather(client::Client, ops::Array{Dispatcher.Op})

Gather the results of all ops. Requires there to be at least one worker available to the scheduler or hangs indefinetely waiting for the results.

source

DaskDistributedDispatcher.shutdown — Method.

shutdown(client::Client)

Tell the dask-scheduler to terminate idle workers and that this client is shutting down. Does NOT terminate the scheduler itself. This does not have to be called after a session but is useful when you want to delete all the information submitted by the client from the scheduler and workers (such as between test runs). If you want to reconnect to the scheduler after calling this function you will have to set up a new client.

source

DaskDistributedDispatcher.default_client — Method.

default_client()

Return the default global client if a client has been registered with the dask-scheduler.

source

DaskDistributedDispatcher.get_key — Method.

get_key(op::Dispatcher.Op)

Calculate an identifying key for op. Keys are re-used for identical ops to avoid unnecessary computations.

source

Worker

DaskDistributedDispatcher.Worker — Type.

Worker

A Worker represents a worker endpoint in the distributed cluster that accepts instructions from the scheduler, fetches dependencies, executes compuations, stores data, and communicates state to the scheduler.

Fields

Communication Management

scheduler_address::Address: the dask-distributed scheduler ip address and port information

batched_stream::Nullable{BatchedSend}: batched stream for communication with scheduler

scheduler::Rpc: manager for discrete send/receive open connections to the scheduler

target_message_size::AbstractFloat: target message size for messages

connection_pool::ConnectionPool: manages connections to peers

total_connections::Integer: maximum number of concurrent connections allowed

Server

address::Address:: ip address and port that this worker is listening on

listener::Base.TCPServer: tcp server that listens for incoming connections

comms::Dict{TCPSocket, String}: current accepted connections to this worker

Handlers

handlers::Dict{String, Function}: handlers for operations requested by open connections

compute_stream_handlers::Dict{String, Function}: handlers for compute stream operations

Data management

data::Dict{String, Any}: maps keys to the results of function calls (actual values)

futures::Dict{String, DeferredFutures.DeferredFuture}: maps keys to their DeferredFuture

nbytes::Dict{String, Integer}: maps keys to the size of their data

types::Dict{String, Type}: maps keys to the type of their data

Task management

tasks::Dict{String, Tuple}: maps keys to the function, args, and kwargs of a task

task_state::Dict{String, String}: maps keys tot heir state: (waiting, executing, memory)

priorities::Dict{String, Tuple}: run time order priority of a key given by the scheduler

priority_counter::Integer: used to also prioritize tasks by their order of arrival

Task state management

transitions::Dict{Tuple, Function}: valid transitions that a task can make

data_needed::Deque{String}: keys whose data we still lack

ready::PriorityQueue{String, Tuple, Base.Order.ForwardOrdering}: keys ready to run

executing::Set{String}: keys that are currently executing

Dependency management

dep_transitions::Dict{Tuple, Function}: valid transitions that a dependency can make

dep_state::Dict{String, String}: maps dependencies with their state (waiting, flight, memory)

dependencies::Dict{String, Set}: maps a key to the data it needs to run

dependents::Dict{String, Set}: maps a dependency to the keys that use it

waiting_for_data::Dict{String, Set}: maps a key to the data it needs that we don't have

pending_data_per_worker::DefaultDict{String, Deque}: data per worker that we want

who_has::Dict{String, Set}: maps keys to the workers believed to have their data

has_what::DefaultDict{String, Set{String}}: maps workers to the data they have

Peer communication

in_flight_tasks::Dict{String, String}: maps a dependency and the peer connection for it

in_flight_workers::Dict{String, Set}: workers from which we are getting data from

suspicious_deps::DefaultDict{String, Integer}: number of times a dependency has not been where it is expected

missing_dep_flight::Set{String}: missing dependencies

Informational

status::String: status of this worker

exceptions::Dict{String, String}: maps erred keys to the exception thrown while running

tracebacks::Dict{String, String}: maps erred keys to the exception's traceback thrown

startstops::DefaultDict{String, Array}: logs of transfer, load, and compute times

Validation

validate::Bool: decides if the worker validates its state during execution

source

DaskDistributedDispatcher.Worker — Method.

Worker(scheduler_address::String; validate=true)

Create a Worker that listens on a random port between 1024 and 9000 for incoming messages. Set validate to false to improve performance.

source

DaskDistributedDispatcher.address — Method.

address(server::Server) -> String

Return this server's address formatted as an String

source

Base.show — Method.

show(io::IO, worker::Worker)

Print a representation of the worker and it's state.

source

Server

DaskDistributedDispatcher.Server — Type.

Server

Abstract type to listen for and handle incoming messages.

source

DaskDistributedDispatcher.address — Function.

address(server::Server) -> String

Return this server's address formatted as an String

source

DaskDistributedDispatcher.start_listening — Function.

listen(server::Server)

Listen for incoming connections on a port and dispatches them to be handled.

source

Address

DaskDistributedDispatcher.Address — Type.

Address

A representation of an endpoint that can be connected to. It is categorized by its scheme (tcp is currently the only protocol supported), host, and port.

source

DaskDistributedDispatcher.Address — Method.

Address(address::String) -> Address

Parse address and returns the corresponding Address object.

source

DaskDistributedDispatcher.Address — Method.

Address(address::String) -> Address

Return the corresponding Address object to the components host and port. By default the tcp protocol is assumed.

source