API

API

Client

DaskDistributedDispatcher.Client — Type.

Client

A Client represents a client that the user can interact with to submit computations to the scheduler and gather results.

source

DaskDistributedDispatcher.Client — Method.

Client(scheduler_address::String) -> Client

Construct a Client which can then be used to submit computations or gather results from the dask-scheduler process.

source

DaskDistributedDispatcher.submit — Method.

submit(client::Client, op::Dispatcher.Op; workers::Array=[])

Submit the Op computation unit to the dask-scheduler for computation.

source

DaskDistributedDispatcher.result — Method.

result(client::Client, op::Dispatcher.Op) -> Any

Gather the result of the Op computation unit. Requires there to be at least one worker available to the scheduler or hangs indefinetely.

source

DaskDistributedDispatcher.cancel — Method.

cancel(client::Client, ops::Array{Dispatcher.Op})

Cancel all Ops in ops. This stops future tasks from being scheduled if they have not yet run and deletes them if they have already run. After calling, this result and all dependent results will no longer be accessible.

source

DaskDistributedDispatcher.gather — Method.

gather(client::Client, ops::Array{Dispatcher.Op})

Gather the results of all ops. Requires there to be at least one worker available to the scheduler or hangs indefinetely waiting for the results.

source

DaskDistributedDispatcher.shutdown — Method.

shutdown(client::Client)

Tell the dask-scheduler to close all workers and that this client is shutting down.

source

DaskDistributedDispatcher.default_client — Method.

default_client()

Return the default global client if a client has been registered with the dask-scheduler.

source

Worker

DaskDistributedDispatcher.Worker — Type.

Worker

A Worker represents a worker endpoint in the distributed cluster that accepts instructions from the scheduler, fetches dependencies, executes compuations, stores data, and communicates state to the scheduler.

source

DaskDistributedDispatcher.Worker — Method.

Worker(scheduler_address::String)

Creates a Worker that listens on a random port for incoming messages.

source

DaskDistributedDispatcher.address — Method.

address(worker::Worker)

Returns this Workers's address formatted as an URI.

source

Base.show — Method.

show(io::IO, worker::Worker)

Prints a representation of the worker and it's state.

source